add workaround for tracing log output via build operation progress for composite builds
build operations of composite settings evaluation is not part of the serialized trace; likely due to build session scope of the trace listener.
This causes problems when parsing log and referecing progress output that have a build operation that is not tracked in the listener. this workaround
just ignores progress events for untraced build operations